better abstraction and moved IntSPODictionary to separate file
